Extracted PropertyBuilderExtensions to file.
